Account for the milestone query table not existing when finding current count
If running for the first time (i.e. this is your first import) the ecatalogue table doesn't exist
and therefore getting a count from it fails. In this scenario we now return a count of 0 and rollback
the connection to allow future SQL statements to run successfully down the same connection.