feat: remove versioning from source records handled by this lib + a few other bits
Sorry, this is a mixed commit, my bad. VersionedRecord is now SourceRecord as we don't need to know the version. There are now a set of utility functions on the SourceRecord too. Additionally, the date associated with an EMu dump is now a date, not a date time/version. And finally, plyvel DB classes have been added. Tests for all too!