feat: add a way to redact records
This isn't _ideal_ but it's the minimum required to allow records to be redacted (this basically means deleted forever). What redacting a record means at the moment is that the ID for the record is added to a redaction database, the record's data is deleted from the backing dataDB, and the record's data will never be added back into the dataDB and therefore views etc ever again. Ideally there would be code in this codebase which also handles the record redactions in MongoDB and Elasticsearch, but I tried coding this up and it was a web or horrors. For redactions of whole records in databases with views that have corresponding Splitgill databases (i.e. specimen, index lot, artefact, and mss currently) this is easy, you just delete the entire MongoDB doc and erase all docs from Elasticsearch. However, things get far more complicated when you attempt to remove non-Splitgill databased views (image, taxonomy) as you have to be able to edit each version of the MongoDB document to remove the offending data (e.g. an image in the associatedImages field) and then put the updated doc back in Mongo, cascading the changes to Elasticsearch as well. For taxonomy, the picture gets murkier as the taxonomy record is merged into the index lot and specimen records, meaning redacting only the taxonomy data is impossible. This requires human thought to decide what is best to do in this scenario as all we can do safely is be incredibly conservative and just delete all fields that could have been merged from the taxonomy record for all versions of the specimen/index lot record. This is most likely overkill but we don't know and have no way to tell without complicated configuration of the redaction. Due to how regularly redactions occur this seemed like it was not worth implementing in code as it is fairly easy to do manually in MongoDB/Elasticsearch (this may actually be fiddly but we I'll have a think if there's some tools that could be created to make this easier). Ultimately doing automatic redacting of records and specific fields is tricky and has a human choice element to it which makes it much better left out of this code (at least of the moment).