fix: fix essentially a buffer overload bug in the DataDB
Found a hard to find bug and this is the fix. The issue arose because the unpacker object in the DataDB is being reused by multiple functions which were generators. The way the msgpack Unpacker object works is that you feed in a bunch of bytes data and then iterate over it to read out the Python objects represented. If the generator feeds 1000 objects worth of raw data into the shared unpacker object and then only reads the first 5 objects out then the unpacker still has 995 objects ready to be read. So when another function comes along and feeds more data and starts reading objects out of the shared unpacker it gets the remaining 995 objects the previous function added to the unpacker. This has been fixed by not using a shared unpacker (we used a shared one for performance but really it was premature optimisation and created a hard to find bug which bamboozled me for the best part of a day!).