feat: remove image orientation field, apply it instead
Instead of adding the orientation tag value to the record, just apply it to the width and height before ingesting. What this means in reality is that the width and height are swapped if the appropriate orientation tag value is found. This is a better approach because it ensures any searches on images are done with the post-orientation values, not pre-orientation so everything is consistent.

Closes: #53