feat: add controlled locking to ensure one importer is running at once
LevelDB implements locking on its databases to ensure cross-process isn't happening, but that results in nasty exceptions which are confusing. We could wrap them, but doing that means we're relying on LevelDB raising this error consistently which feels a bit meh, better to just have our own lock in place and check that. This commit has a new dependency for filelock which is used to create a file in the dimp data directory which controls access.

Closes: #52