feat!: split is_member into is_member and is_publishable
This splits the rules defined in the is_member method of view classes into two methods: is_member contains the rules that define fundamental membership of a view, e.g. is it a "Specimen" record type, and is_publishable contains the rules that define whether it should currently be published, e.g. is it on loan. This will of course increase the number of records processed by the importer at certain stages (more records are considered members) but it's necessary so that we're not accidentally publishing records that we shouldn't be.

BREAKING CHANGE: split is_member into is_member and is_publishable

Closes: #57