feat: add importer method to purge unsuitable records
A small number of records may change their properties so that they are no longer a member of the view or so that they can no longer be published under the rules of the view. The latter case should be handled by the regular ingest process (see PR #60), but records that are no longer members are not. This can be run occasionally to fix both issues.